requirements.txt
aniso8601==9.0.1
api==0.0.7
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.10.0
black==21.12b0
bs4==0.0.1
CacheControl==0.12.11
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
coverage==6.3.3
crashtest==0.3.1
cryptography==37.0.2
distlib==0.3.4
Django==4.0.1
django-nose==1.4.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
factory-boy==3.2.1
Faker==11.3.0
filelock==3.7.1
Flask==2.1.1
Flask-Cors==3.0.10
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
html5lib==1.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
jwt==1.3.1
keyring==23.6.0
lockfile==0.12.2
MarkupSafe==2.1.1
marshmallow==3.14.1
marshmallow-dataclass==8.5.8
msgpack==1.0.4
mypy==0.961
mypy-extensions==0.4.3
nose==1.3.7
packaging==20.9
parameterized==0.8.1
pastel==0.2.1
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.2.0
pinocchio==0.4.3
pkginfo==1.8.3
platformdirs==2.4.1
pluggy==1.0.0
poetry==1.1.14
poetry-core==1.0.8
psycopg2==2.9.3
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pylev==1.4.0
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.27.1
requests-toolbelt==0.9.1
router==0.1
shellingham==1.4.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlparse==0.4.2
statistic==0.1.6
text-unidecode==1.3
tomli==1.2.3
tomlkit==0.11.1
typing-inspect==0.7.1
typing_extensions==4.0.1
tzdata==2022.1
urllib3==1.26.9
views==0.3
virtualenv==20.15.1
webencodings==0.5.1
Werkzeug==2.1.1
pytest_factoryboy==2.5.0
